Patch to QB64.bas to fix minor IDE display glitch
Since the clock works as an ON TIMER event, it could try to print to the bottom of the IDE at the wrong time (such as while resizing and such), which would cause an IDE module error.  This patch corrects that so it'll skip printing a seconds update, unless it can do so safely without screwing up the LOCATE coordinates.